find

+++ Agenda.js
@@ -22,11 +22,11 @@ export default class Agenda {
         let cellFecha = row.insertCell(1);
         let cellDias = row.insertCell(2);
         row.insertCell(3);
-   
+
         cellNombre.innerHTML = tarea.nombre;
         cellFecha.innerHTML = tarea.getFecha();
         cellDias.innerHTML = tarea.getDias();
-        this._borrar(row,tarea);
+        this._borrar(row, tarea);
 
         let objTarea = {
             nombre: tarea.nombre,
@@ -35,11 +35,23 @@ export default class Agenda {
         this._tareas.push(objTarea);
     }
 
+    //Agregar tarea
     addTarea(tarea) {
+        let found = this._findTarea(tarea.nombre);
+        if (found >= 0) {
+            Swal.fire({
+                type: "error",
+                tittle: "Error",
+                text: "La tarea ya existe"
+            });
+            return;
+        }
         this._showInTable(tarea);
         localStorage.setItem("tareas", JSON.stringify(this._tareas));
     }
 
+
+    //Borrar fila//
     _borrar(row, tarea) {
         let btnBorrar = document.createElement("input");
         btnBorrar.type = "button";
@@ -49,7 +61,7 @@ export default class Agenda {
         row.cells[3].appendChild(btnBorrar);
         btnBorrar.addEventListener("click", () => {
             this._borrarRow(tarea);
-            
+
         });
     }
 
@@ -63,7 +75,7 @@ export default class Agenda {
         location.reload();
         localStorage.setItem("tareas", JSON.stringify(this._tareas));
     }
-
+    //Ordenar alfabeticamente//
     _compararAlfa(a, b) {
         if (a.nombre < b.nombre) {
             return -1;
@@ -81,11 +93,41 @@ export default class Agenda {
         localStorage.setItem("tareas", JSON.stringify(this._tareas));
         location.reload();
     }
+    //Ordenar numericamente//
+    _compararNum(a, b) {
+        if (a.fecha < b.fecha) {
+            return -1;
+        }
+        if (a.fecha > b.fecha) {
+            return 1;
+        }
+        return 0;
+    }
+    _num() {
+        this._tareas.sort(this._compararNum);
+    }
+    mostrarNum() {
+        this._tareas.sort(this._compararNum);
+        localStorage.setItem("tareas", JSON.stringify(this._tareas));
+        location.reload();
+    }
+
+    //Encontrar tarea
+    _findTarea(nombre) {
+        let foundAt = -1;
+        this._tareas.forEach((e, index) => {
+            if (e.nombre === nombre) {
+                foundAt = index;
+                return;
+            }
+        });
+        return foundAt;
+    }
 
 
 
 
 
-   
-   
+
+
 }
